+++ client/components/NavBar/Settings.jsx
