+++ client/components/auth-components/OTP.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useContext, useEffect } from 'react';
-import { RecoveryContext } from '../App';
+import { RecoveryContext } from '../../App';
 import { useNavigate } from 'react-router';
 
 const OTP = () => {
